+++ transform/apop_fix_params.c
@@ -121,13 +121,13 @@ Apop_settings_init(apop_fix_params,
 Apop_settings_copy(apop_fix_params, )
 Apop_settings_free(apop_fix_params, )
 
-static double fix_params_ll(apop_data *d, apop_model *fixed_model){
+static long double fix_params_ll(apop_data *d, apop_model *fixed_model){
     apop_model *base_model = Apop_settings_get(fixed_model, apop_fix_params, base_model);
     unpack(base_model->parameters, fixed_model);
     return apop_log_likelihood(d, base_model);
 }
 
-static double fix_params_p(apop_data *d, apop_model *fixed_model){
+static long double fix_params_p(apop_data *d, apop_model *fixed_model){
     apop_model *base_model = Apop_settings_get(fixed_model, apop_fix_params, base_model);
     unpack(base_model->parameters, fixed_model);
     return apop_p(d, base_model);
